Add files via upload
After creating the navigation bar, I added the TrackZen logo and added the home, dashboard, analytic, help and login/signup button. I added a drop down in the analytic and added 4 buttons called screenshots, network, device info, and report. All those buttons were made to pop up.I created the login form and put two boxes in it as enter your email and enter your password and put the forgot password option and put the login button as a pop up button.
Don't have an account and put the signup option in it says already have an account and linked the login form there. When that button is opened, go to the signup form and enter your email, create password and confirm your password in the signup form and put the sgnup button and put that button as a pop up button. Underneath, it says already have an account and linked the login form there